Android: Bugfixes, Cache Control, Refactoring for Story Grid

+++ clients/android/NewsBlur/src/com/newsblur/util/FileCache.java
@@ -15,13 +15,11 @@ public class FileCache {
     private static final long MIN_FREE_SPACE_BYTES = 250L * 1024L * 1024L;
     private static final Pattern POSTFIX_PATTERN = Pattern.compile("(\\.[a-zA-Z0-9]+)[^\\.]*$");
 
-    private final long maxFileAgeMillis;
     private final int minValidCacheBytes;
 
 	private final File cacheDir;
 
-	private FileCache(Context context, String subdir, long maxFileAgeMillis, int minValidCacheBytes) {
-        this.maxFileAgeMillis = maxFileAgeMillis;
+	private FileCache(Context context, String subdir, int minValidCacheBytes) {
         this.minValidCacheBytes = minValidCacheBytes;
         cacheDir = new File(context.getCacheDir(), subdir);
 		if (!cacheDir.exists()) {
@@ -30,17 +28,17 @@ public class FileCache {
 	}
 
     public static FileCache asStoryImageCache(Context context) {
-        FileCache fc = new FileCache(context, "olimages", 30L * 24L * 60L * 60L * 1000L, 512);
+        FileCache fc = new FileCache(context, "olimages", 512);
         return fc;
     }
 
     public static FileCache asIconCache(Context context) {
-        FileCache fc = new FileCache(context, "icons", 45L * 24L * 60L * 60L * 1000L, 128);
+        FileCache fc = new FileCache(context, "icons", 128);
         return fc;
     }
 
     public static FileCache asThumbnailCache(Context context) {
-        FileCache fc = new FileCache(context, "thumbs", 15L * 24L * 60L * 60L * 1000L, 256);
+        FileCache fc = new FileCache(context, "thumbs", 256);
         return fc;
     }
 
@@ -109,22 +107,29 @@ public class FileCache {
         return fileName;
     }
 
-    public void cleanupOld() {
+    public void cleanupOld(long maxFileAgeMillis) {
         try {
+            int cleaned = 0;
             File[] files = cacheDir.listFiles();
             if (files == null) return;
+            com.newsblur.util.Log.i(this, String.format( "have %d files", files.length));
             for (File f : files) {
                 long timestamp = f.lastModified();
                 if (System.currentTimeMillis() > (timestamp + maxFileAgeMillis)) {
                     f.delete();
+                    cleaned++;
                 }
             }
+            com.newsblur.util.Log.i(this, String.format( "cleaned up %d files", cleaned));
         } catch (Exception e) {
-            android.util.Log.e(FileCache.class.getName(), "exception cleaning up cache", e);
+            com.newsblur.util.Log.e(this, "exception cleaning up cache", e);
         }
     }
 
-    public void cleanupUnusedOrOld(Set<String> currentUrls) {
+    /**
+     * Clean up files in this cache that are both unused and past the specified age.
+     */
+    public void cleanupUnusedAndOld(Set<String> currentUrls, long maxFileAgeMillis) {
         // if there appear to be zero images in the system, a DB rebuild probably just
         // occured, so don't trust that data for cleanup
         if (currentUrls.size() == 0) return;
@@ -132,17 +137,21 @@ public class FileCache {
         Set<String> currentFiles = new HashSet<String>(currentUrls.size());
         for (String url : currentUrls) currentFiles.add(getFileName(url));
         try {
+            int cleaned = 0;
             File[] files = cacheDir.listFiles();
             if (files == null) return;
+            com.newsblur.util.Log.i(this, String.format( "have %d files", files.length));
             for (File f : files) {
                 long timestamp = f.lastModified();
-                if ((System.currentTimeMillis() > (timestamp + maxFileAgeMillis)) ||
+                if ((System.currentTimeMillis() > (timestamp + maxFileAgeMillis)) &&
                     (!currentFiles.contains(f.getName()))) {
                     f.delete();
+                    cleaned++;
                 }
             }
+            com.newsblur.util.Log.i(this, String.format( "cleaned up %d files", cleaned));
         } catch (Exception e) {
-            android.util.Log.e(FileCache.class.getName(), "exception cleaning up cache", e);
+            com.newsblur.util.Log.e(this, "exception cleaning up cache", e);
         }
     }
